Upgrade config to use Berkshelf and ruby-install
This splits the chef configuration into different recipes that support
building the components separately.  A worker/server node, app, db, and
search server can be built, or the "development" recipe will build all
the them on one box.

The only part that's tested fully is the development and worker
node recipes.